+++ extensions/data-plane-selector/selector-client/src/main/java/org/eclipse/dataspaceconnector/dataplane/selector/DataPlaneInstanceClientExtension.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.dataplane.selector;
 
-import net.jodah.failsafe.RetryPolicy;
+import dev.failsafe.RetryPolicy;
 import okhttp3.OkHttpClient;
 import org.eclipse.dataspaceconnector.common.string.StringUtils;
 import org.eclipse.dataspaceconnector.dataplane.selector.client.DataPlaneSelectorClient;